mailer code pulled out of index.js file
