+++ api/index.js
@@ -1,15 +1,15 @@
 import express from 'express';
-import nodemailer from'nodemailer';
+import { setMailOptions } from './mailer';
 
-var emailPassword = process.env.EMAILPASSWORD;
-
-var transporter = nodemailer.createTransport({
-  service: 'gmail',
-  auth: {
-    user: 'teamglow2017@gmail.com',
-    pass: emailPassword
-  }
-});
+// var emailPassword = process.env.EMAILPASSWORD;
+//
+// var transporter = nodemailer.createTransport({
+//   service: 'gmail',
+//   auth: {
+//     user: 'teamglow2017@gmail.com',
+//     pass: emailPassword
+//   }
+// });
 
 const router = express.Router();
 var Like = require('../models/likes.js');
@@ -77,37 +77,37 @@ router.post('/users/:username/likes', (req, res)=>{
   });
 });
 
-var setMailOptions =  function (like_id) {
-  Like.find({ _id: like_id }).then(like =>{
-    var likeInfo = like[0];
-    User.find({username: likeInfo.usernameOfLiked_id}).then(user =>{
-      const mailOptions = { from: 'teamglow2017@gmail.com', // sender address
-        to: `${user[0].email}`, // list of receivers
-        subject: 'You have been liked....', // Subject line
-        html: `<p>You have been liked today by ${likeInfo.nameOfLiker}</p>
-       <br>${likeInfo.nameOfLiker} asked us to pass on this message:
-       <br>
-       <br>${likeInfo.message}.
-       <br>
-       <br>To get in touch you can email ${likeInfo.nameOfLiker} on ${likeInfo.emailOfLiker}.
-       <br>
-       <br>Have fun getting your geek on...
-       <br>
-       <br>The Dind'r Team`};
-       // plain text body
-      sendEmailNotification(mailOptions);
-    });
-  });
-};
+// var setMailOptions =  function (like_id) {
+//   Like.find({ _id: like_id }).then(like =>{
+//     var likeInfo = like[0];
+//     User.find({username: likeInfo.usernameOfLiked_id}).then(user =>{
+//       const mailOptions = { from: 'teamglow2017@gmail.com', // sender address
+//         to: `${user[0].email}`, // list of receivers
+//         subject: 'You have been liked....', // Subject line
+//         html: `<p>You have been liked today by ${likeInfo.nameOfLiker}</p>
+//        <br>${likeInfo.nameOfLiker} asked us to pass on this message:
+//        <br>
+//        <br>${likeInfo.message}.
+//        <br>
+//        <br>To get in touch you can email ${likeInfo.nameOfLiker} on ${likeInfo.emailOfLiker}.
+//        <br>
+//        <br>Have fun getting your geek on...
+//        <br>
+//        <br>The Dind'r Team`};
+//        // plain text body
+//       sendEmailNotification(mailOptions);
+//     });
+//   });
+// // };
+// var sendEmailNotification = function(mailOptions) {
+//   transporter.sendMail(mailOptions, function (err, info){
+//     if (err)
+//       console.log(err);
+//     else
+//     console.log(info);
+//   });
+// };
 
-var sendEmailNotification = function(mailOptions) {
-  transporter.sendMail(mailOptions, function (err, info){
-    if (err)
-      console.log(err);
-    else
-    console.log(info);
-  });
-};
 
 var findUserLikes = function(req,res){
   var username = req.params.username;
