+++ server.js
@@ -4,6 +4,7 @@ import apiRouter from './api';
 import bodyParser from 'body-parser';
 import mongoose from 'mongoose';
 import profile from './api/profile';
+import users from './api/users';
 
 var url = process.env.MONGOLAB_URI;
